initial plugins changes

+++ plugins/localfile/localfile.go
@@ -22,7 +22,7 @@ type Plugin struct {
 	FilePath string
 	Logger   *logrus.Logger
 	hostname string
-	interval int
+	interval float64
 }
 
 // Delimiter defines what kind of delimiter we'll use in the CSV format -- in this case, we want TSV
@@ -40,7 +40,7 @@ func (p *Plugin) Flush(ctx context.Context, metrics []samplers.InterMetric) erro
 	return nil
 }
 
-func appendToWriter(appender io.Writer, metrics []samplers.InterMetric, hostname string, interval int) error {
+func appendToWriter(appender io.Writer, metrics []samplers.InterMetric, hostname string, interval float64) error {
 	gzW := gzip.NewWriter(appender)
 	csvW := csv.NewWriter(gzW)
 	csvW.Comma = Delimiter
